+++ CHANGES.md
@@ -34,6 +34,7 @@ UI:
 -   Added UI for "Link to an API or web service" box
 -   Added UI for storing files
 -   Added UI for "Publish as Open Data to data.gov.au"
+-   Refactor view dataset page & remove the edit function from the page
 -   Clarify tooltip text
 
 Storage: